requirements.txt
pyyaml
pandas
smbus2
Phidget22

